Add multi-line input support
The way this multi-line input works is very much inspired by the
PureScript REPL which also uses `haskeline`: https://github.com/purescript/purescript/issues/934#issuecomment-80525135

The name of the command is configurable, so it could be called `m`,
`paste`, `multiline`, or whatever works best for the application using
this.

The banner can now take into account if it's in the middle of a
multi-line input or not.